trace/requirements.txt
fastapi
uvicorn
httpx[http2]
kafka-python